Update and rename Applied_AI_MB_Code.ipynb to Multi-Arm Bandit Optimization: Comparing Epsilon-Greedy Strategies.ipynb
This project explores the concept of multi-arm bandit problems and compares the performance of different epsilon-greedy strategies. The code implements a class called "multi_arm_bandit" that simulates the bandit problem and includes methods for epsilon-greedy action selection. The project investigates the effects of different epsilon values on the average reward obtained over a specified number of steps.