Page Replacement Algorithms Simulator.py
The provided code simulates two-page replacement algorithms: FIFO (First-In-First-Out) and LFU (Least Frequently Used). The project can be named "Page Replacement Algorithms Simulator" as it focuses on simulating and comparing these two algorithms.

The code maintains a list called "request" to store the page numbers and a list called "cache" to represent the memory cache. It includes the implementation of the FIFO and LFU algorithms using the provided functions.

The user is prompted to enter page numbers until they input 0 to indicate the end of the page requests. Then, the user can choose between FIFO (option 1) or LFU (option 2) algorithms. The code simulates the selected algorithm, prints whether a page request results in a "Hit" or a "Miss," and updates the cache accordingly.

The project allows for comparing the performance of these page replacement algorithms based on the cache hits and misses.